requirements.txt
flask==2.0.1
docker==5.0.0
gitpython==3.1.24
psutil==5.8.0
werkzeug==2.0.1